centralize AIFormMessage and add disclaimer
