+++ quadratic-client/src/app/ui/menus/CodeEditor/AIAssistant/AIAssistant.tsx
@@ -1,3 +1,4 @@
+import { AIUserMessageFormDisclaimer } from '@/app/ui/components/AIUserMessageForm';
 import { AIAssistantMessages } from '@/app/ui/menus/CodeEditor/AIAssistant/AIAssistantMessages';
 import { AIAssistantUserMessageForm } from '@/app/ui/menus/CodeEditor/AIAssistant/AIAssistantUserMessageForm';
 import { useRef } from 'react';
@@ -9,7 +10,10 @@ export const AIAssistant = ({ autoFocus }: { autoFocus?: boolean }) => {
     <div className="grid h-full grid-rows-[1fr_auto]">
       <AIAssistantMessages textareaRef={textareaRef} />
 
-      <AIAssistantUserMessageForm ref={textareaRef} autoFocus={autoFocus} textareaRef={textareaRef} />
+      <div className="flex h-full flex-col justify-end px-2 py-0.5">
+        <AIAssistantUserMessageForm ref={textareaRef} autoFocus={autoFocus} textareaRef={textareaRef} />
+        <AIUserMessageFormDisclaimer />
+      </div>
     </div>
   );
 };
